implement move left, and block list
In order to test if moving left is legit, I also had to add the list of
blocks and their rotated forms. Moving left will fail (result in the
same state) if the active block is already up against the wall or if it
would overlap with a previously fallen block - or in other words, if it
would result in an illegal state.

Getting this infrastructure out of the way should make most of the other
actions pretty straightforward!